// third-party/ketcher/ketcher-react/src/ui/action/action.types.ts
type ActionObj = {
  tool?: string;
  opts?: unknown;
  dialog?: string;
  thunk?: (dispatch: Dispatch, getState: () => ActionThunkState) => void;
};
type ActionFn = (editor: ActionEditorContext) => void;
// todo: come up with better name
type UiActionAction = ActionObj | ActionFn;

// todo: come up with better name
type GetSelectedState = (
  editor: ActionEditorContext,
  server?: unknown,
) => boolean;
type GetDisabledState = (
  editor: ActionEditorContext,
  server: unknown,
  options: ActionStateOptions,
) => boolean;
type GetHiddenState = (options: ActionStateOptions) => boolean;

export type GetActionState =
  | GetSelectedState
  | GetDisabledState
  | GetHiddenState;

type IsSelectedState = boolean | GetSelectedState;
type IsDisabledState = boolean | GetDisabledState;
type IsHiddenState = boolean | GetHiddenState;

interface UiAction {
  title?: string;
  shortcut?: string | Array<string>;
  enabledInViewOnly?: true;
  action: UiActionAction;
  selected?: IsSelectedState;
  disabled?: IsDisabledState;
  hidden?: IsHiddenState;
  onAction?: (action: UiActionAction) => void;
}